promote DelegatingTimelockService
